requirements.txt
python-docx
PyPDF2
attrs
cachetools
click
cryptography
flask
Flask-Cors
Flask-JWT-Extended
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
greenlet
gunicorn
httplib2
importlib-metadata
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
numpy
oauthlib
packaging
pandas
pkgutil-resolve-name
protobuf
pyasn1
pyasn1-modules
PyJWT
pyparsing
python-dateutil
referencing
requests
requests-oauthlib
SQLAlchemy
typing-extensions
tzdata
uritemplate
urllib3
werkzeug
zipp
python-dotenv
openai==0.28
openpyxl
psycopg2-binary
anthropic
pydantic
mongoengine
cffi
alembic